main() moved into Application
